refactor: add chat analysis cmds file
